requirements.txt
Flask==2.1.2
google-api-core==2.7.3
google-auth==2.6.6
google-cloud-texttospeech==2.11.0
googleapis-common-protos==1.56.0


pandas~=1.4.2
openpyxl~=3.0.9